Users that adopt a pet can edit pet information. Update dog profile picture

+++ database/pets_profile.php
@@ -1,4 +1,26 @@
 <?php
+    /*
+    * Check if user has updated permissions
+    */
+    function canUpdate($idUser, $idPet) {
+        $db = Database::instance()->db();
+        
+        $owner = $db->prepare('SELECT * FROM UserFoundPet WHERE idUser = ? and idPet = ?');
+        $owner->execute(array($idUser, $idPet));
+        $isOwner = $owner->fetchAll();
+
+        $adopted = $db->prepare('SELECT * FROM UserAdoptedPet WHERE idUser = ? and idPet = ?');
+        $adopted->execute(array($idUser, $idPet));
+        $isAdopted = $adopted->fetchAll();
+        
+        if(empty($isOwner) && empty($isAdopted)) {
+            return FALSE;
+        } 
+        else {
+            return TRUE;
+        }
+    }
+
     /* 
      * Update and Add a new Pet
     */
@@ -25,22 +47,37 @@
         return TRUE;
     }
 
-    function updatePet($idPet, $npetName, $bio, $nspecie, $ngender, $nsize, $ncolor) {
+    function updatePet($idUser, $idPet, $npetName, $bio, $nspecie, $ngender, $nsize, $ncolor) {
         $db = Database::instance()->db();
         
-        $stmt = $db->prepare('UPDATE Pet SET petName = ?, bio = ?, specie = ?, gender = ?, size = ?, color = ? WHERE idPet = ?');
+        if(canUpdate($idUser, $idPet)) {
+            $stmt = $db->prepare('UPDATE Pet SET petName = ?, bio = ?, specie = ?, gender = ?, size = ?, color = ? WHERE idPet = ?');
 
-        $hashed_new_password = sha1($new_password);
-        $stmt->execute(array( $npetName, $bio, $nspecie, $ngender, $nsize, $ncolor, $idPet));  
+            $hashed_new_password = sha1($new_password);
+            $stmt->execute(array($npetName, $bio, $nspecie, $ngender, $nsize, $ncolor, $idPet));  
 
+            $originalFileName = "../images/pet-profile/pet-$idPet/profile.jpg";
+            move_uploaded_file($_FILES['image']['tmp_name'], $originalFileName);
 
-        $originalFileName = "../images/pet-profile/pet-$idPet/profile.jpg";
-        move_uploaded_file($_FILES['image']['tmp_name'], $originalFileName);
+            return TRUE;
+        }
+
+        return FALSE;
     }
 
     /*
     * Pet Photos
     */
+    function getPetFromPhoto($idPhoto) {
+        $db = Database::instance()->db();
+        
+        $stmt = $db->prepare('SELECT * FROM Pet, PetPhoto WHERE idPhoto = ? and PetPhoto.idPet = Pet.idPet');
+        $stmt->execute(array($idPhoto));
+        $pet = $stmt->fetch();
+
+        return $pet;
+    }
+
     function getAllPhotos($id) {
         $db = Database::instance()->db();
         
@@ -52,33 +89,41 @@
         return $petsID;
     }
 
-    function addPetPhoto($idPet) {
+    function addPetPhoto($idUser, $idPet) {
         $db = Database::instance()->db();
         
-        $stmt = $db->prepare('INSERT INTO PetPhoto(idPet) VALUES (?)');
-        $stmt->execute(array($idPet));
+        if(canUpdate($idUser, $idPet)) {
+            $stmt = $db->prepare('INSERT INTO PetPhoto(idPet) VALUES (?)');
+            $stmt->execute(array($idPet));
 
-        $idPhoto = $db->lastInsertId();
+            $idPhoto = $db->lastInsertId();
 
-        if (!file_exists("../images/pet-profile/pet-$idPet")) {
-            mkdir("../images/pet-profile/pet-$idPet", 0777, true);
+            if (!file_exists("../images/pet-profile/pet-$idPet")) {
+                mkdir("../images/pet-profile/pet-$idPet", 0777, true);
+            }
+
+            $originalFileName = "../images/pet-profile/pet-$idPet/photo-$idPhoto.jpg";
+
+            // Move the uploaded file to its final destination
+            move_uploaded_file($_FILES['image-album']['tmp_name'], $originalFileName);
+
+            return TRUE;
         }
-
-        $originalFileName = "../images/pet-profile/pet-$idPet/photo-$idPhoto.jpg";
-
-        // Move the uploaded file to its final destination
-        move_uploaded_file($_FILES['image']['tmp_name'], $originalFileName);
-
-        return TRUE;
+        return FALSE;
     }
 
-    function deletePhoto($idPhoto) {
+    function deletePhoto($idUser, $idPhoto) {
         $db = Database::instance()->db();
         
-        $stmt = $db->prepare('DELETE FROM PetPhoto WHERE idPhoto = ?');
-        $stmt->execute(array($idPhoto));
+        $pet = getPetFromPhoto($idPhoto);
+        if((!empty($pet)) && (canUpdate($idUser, $pet['idPet']))) {
+            $stmt = $db->prepare('DELETE FROM PetPhoto WHERE idPhoto = ?');
+            $stmt->execute(array($idPhoto));
 
-        return TRUE;
+            return TRUE;
+        }
+
+        return FALSE;
     }
 
     /*
@@ -110,18 +155,27 @@
         if(empty($petsID)) {
             $stmt = $db->prepare('INSERT INTO FavoritePet VALUES (?, ?)');
             $stmt->execute(array($user['idUser'], $idPet));
-            print('add');
         }
         else {
             $stmt = $db->prepare('DELETE FROM FavoritePet WHERE idUser = ? and idPet = ?');
             $stmt->execute(array($user['idUser'], $idPet));
-            print('del');
         }
     }
 
     /*
     *   Pet Posts
     */
+    function getPetFromPost($idPost) {
+        $db = Database::instance()->db();
+
+        $stmt = $db->prepare('SELECT * FROM Pet, PostsPet WHERE Pet.idPet = PostsPet.idPet and id = ?');
+        
+        $stmt->execute(array($idPost));
+        $pet = $stmt->fetch();
+
+        return $pet;
+    }
+
     function getPosts($idPet) {
         $db = Database::instance()->db();
 
@@ -133,30 +187,60 @@
         return $petsID;
     }
     
-    function addPost($idPet, $username, $post) {
+    function addPost($user, $idPet, $post) {
         $db = Database::instance()->db();
         
-        $stmt = $db->prepare('INSERT INTO PostsPet(idPet, author, datePost, post) VALUES (?, ?, ?, ?)');
-        $stmt->execute(array($idPet, $username, date("Y/m/d"), $post));
+        if(canUpdate($user['idUser'], $idPet)) {
+            $stmt = $db->prepare('INSERT INTO PostsPet(idPet, author, datePost, post) VALUES (?, ?, ?, ?)');
+            $stmt->execute(array($idPet, $user['username'], date("Y/m/d"), $post));
+
+            return TRUE;
+        }
+
+        return FALSE;
     }
 
-    function updatePost($idPost, $info) {
+    function updatePost($idUser, $idPost, $info) {
         $db = Database::instance()->db();
         
-        $stmt = $db->prepare('UPDATE PostsPet SET POST = ? WHERE id = ?');
-        $stmt->execute(array($info, $idPost));  
+        $pet = getPetFromPost($idPost);
+        if((!empty($pet)) && canUpdate($idUser, $pet['idPet'])) {
+            $stmt = $db->prepare('UPDATE PostsPet SET POST = ? WHERE id = ?');
+            $stmt->execute(array($info, $idPost)); 
+            
+            return TRUE;
+        }
+        return FALSE;
     }
 
-    function deletePost($idPost) {
+    function deletePost($idUser, $idPost) {
         $db = Database::instance()->db();
         
-        $stmt = $db->prepare('DELETE FROM PostsPet WHERE id = ?');
-        $stmt->execute(array($idPost));
+        $pet = getPetFromPost($idPost);
+        
+        if((!empty($pet)) && canUpdate($idUser, $pet['idPet'])) {
+            $stmt = $db->prepare('DELETE FROM PostsPet WHERE id = ?');
+            $stmt->execute(array($idPost));
+
+            return TRUE;
+        }
+
+        return FALSE;
     }
 
     /* 
     * Pet Questions and Answer 
     */
+    function getPetFromQuestion($idQuestion) {
+        $db = Database::instance()->db();
+        
+        $stmt = $db->prepare('SELECT * FROM Pet, PetQuestion WHERE PetQuestion.idQuestion = ? and PetQuestion.idPet = Pet.idPet');
+        $stmt->execute(array($idQuestion));
+        $pet = $stmt->fetch();
+
+        return $pet;
+    }
+
     function getQuestions($idPet) {
         $db = Database::instance()->db();
 
@@ -178,32 +262,28 @@
     function deleteQuestion($idUser, $idQuestion) {
         $db = Database::instance()->db();
 
-        $stmt = $db->prepare('SELECT * FROM User, UserFoundPet, PetQuestion WHERE 
-                                                                            PetQuestion.idQuestion = ? 
-                                                                            and PetQuestion.idPet = UserFoundPet.idPet
-                                                                            and User.idUser = UserFoundPet.idUser');
-        $stmt->execute(array($idQuestion));
-        $question = $stmt->fetch();
-
-        if((!empty($question)) && ($question['idUser'] == $idUser)) {
+        $pet = getPetFromQuestion($idQuestion);
+        if((!empty($pet)) && (canUpdate($idUser, $pet['idPet']))) {
             $stmt = $db->prepare('DELETE FROM PetQuestion WHERE idQuestion = ?');
             $stmt->execute(array($idQuestion));
+
+            return TRUE;
         }
+
+        return FALSE;
     }
 
-    function addAnswer($idQuestion, $author, $answer) {
+    function addAnswer($user, $idQuestion, $answer) {
         $db = Database::instance()->db();
         
-        $stmt = $db->prepare('SELECT * FROM User, UserFoundPet, PetQuestion WHERE 
-                                                                            PetQuestion.idQuestion = ? 
-                                                                            and PetQuestion.idPet = UserFoundPet.idPet
-                                                                            and User.idUser = UserFoundPet.idUser');
-        $stmt->execute(array($idQuestion));
-        $question = $stmt->fetch();
-
-        if((!empty($question)) && ($question['username'] == $author)) {
+        $pet = getPetFromQuestion($idQuestion);
+        if((!empty($pet)) && (canUpdate($user['idUser'], $pet['idPet']))) {
             $update = $db->prepare('UPDATE PetQuestion SET dateAnswer = ?, authorAnswer = ?, answer = ? WHERE idQuestion = ?');
-            $update->execute(array(date("Y/m/d"), $author, $answer, $idQuestion));  
+            $update->execute(array(date("Y/m/d"), $user['username'], $answer, $idQuestion));  
+
+            return TRUE;
         }
+
+        return FALSE;
     }
 ?>
